Update README.md
Swiggy is a popular food delivery platform known for its user-friendly interface and seamless user experience. While it excels in many areas, there are opportunities for enhancement and innovation. Foodggy aims to leverage these opportunities to create a superior food delivery application that provides an improved UI and UX.ObjectiveThe objective is to identify specific areas in the UI and UX of Swiggy that can be improved and implement these improvements in the Foodggy app. The focus will be on enhancing user engagement, simplifying navigation, and providing a more personalized experience.so,in inned of that I have designed the UI for my own app Foodggy.